Move more create process over, remove node dep on terminal instance
